requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.6.0
backcall==0.1.0
blinker==1.4
cachetools==4.0.0
certifi==2018.1.18
cffi==1.12.3
chardet==3.0.4
click==6.7
click-plugins==1.1.1
cligj==0.5.0
cloud-init==18.3
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.7
decorator==4.4.0
distro-info==0.18
Django==2.2.5
django-baton==1.5.7
django-geojson==2.12.0
django-jsonfield==1.3.1
django-leaflet==0.25.0
django-simple-history==2.7.3
et-xmlfile==1.0.1
Fiona==1.8.9
GDAL==2.2.2
geopandas==0.6.1
google-api-python-client==1.7.11
google-auth==1.10.1
google-auth-httplib2==0.0.3
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
ipython==7.8.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.1
jellyfish==0.7.2
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lib50==2.0.0
MarkupSafe==1.0
munch==2.3.2
numpy==1.13.3
oauth2client==1.5.2
oauthlib==2.0.6
openpyxl==2.6.3
PAM==0.4.2
pandas==0.25.2
parso==0.5.1
petl==1.3.0
pexpect==4.7.0
pickleshare==0.7.5
pipenv==2018.11.26
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.19
pycrypto==2.6.1
PyGeoj==1.0.0
Pygments==2.4.2
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyproj==2.4.0
pyserial==3.4
python-apt==1.6.2
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2019.2
pyxdg==0.25
PyYAML==3.12
requests==2.22.0
requests-unixsocket==0.1.5
rsa==4.0
SecretStorage==2.3.1
service-identity==16.0.0
Shapely==1.6.4.post2
six==1.11.0
sqlparse==0.3.0
ssh-import-id==5.7
submit50==3.0.2
systemd-python==234
termcolor==1.1.0
traitlets==4.3.2
Twisted==17.9.0
ufw==0.35
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.22
virtualenv==16.7.6
virtualenv-clone==0.5.3
wcwidth==0.1.7
xlrd==1.2.0
zope.interface==4.3.2
